export game logic to game.js
